BioemePathts.kt can parse and serialize the same way

+++ app/src/commonMain/kotlin/model/Point.kt
@@ -20,6 +20,6 @@
 package model
 
 data class Point(
-    val x: Float,
-    val y: Float
+    val x: Int,
+    val y: Int
 )
